Add convenient method for fire and forget messages
For the common cases where the user simply wanted to create an instance
of something that's `Safe` and immediately send it, the syntax was
really verbose. First, the user had to use `mkBox` or `mkBoxOf` to
create the box, and then populate it with the desired fields (and
somewhere along the way extract the desired permission from the `Packed`
instance), and finally send the message.

The aforementioned process is now replaced with `a >!< msg`, where `msg`
has the same type of restrictions that the argument for `Box.mkBoxOf`
has.